print window cutted out into new py file fixed 4up print mode
